requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
pydantic==2.8.2
